FIX: type of return

+++ 0x00-python_variable_annotations/2-floor.py
@@ -3,6 +3,6 @@
 and returns the floor of the float """
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """ Function that take a argument and returns the floor of the float """
-    return n // 1
+    return int(n // 1)
